+++ apps/site/src/app/admin/layout/AdminLayout.tsx
@@ -8,7 +8,7 @@ import AppLayout from "@cloudscape-design/components/app-layout";
 import axios from "axios";
 import { SWRConfig } from "swr";
 
-import { isAdminRole } from "@/lib/admin/adminRole";
+import { isAdminRole } from "@/lib/admin/authorization";
 import UserContext from "@/lib/admin/UserContext";
 import useUserIdentity from "@/lib/admin/useUserIdentity";
