+++ apps/site/src/app/admin/layout/AdminSidebar.tsx
@@ -1,20 +1,37 @@
 import { usePathname } from "next/navigation";
 
+import { useContext } from "react";
+
 import SideNavigation, {
 	SideNavigationProps,
 } from "@cloudscape-design/components/side-navigation";
 
 import { BASE_PATH, useFollowWithNextLink } from "./common";
+import { isApplicationManager } from "@/lib/admin/authorization";
+import UserContext from "@/lib/admin/UserContext";
 
 function AdminSidebar() {
 	const pathname = usePathname();
 	const followWithNextLink = useFollowWithNextLink();
 
+	const { role } = useContext(UserContext);
+
 	const navigationItems: SideNavigationProps.Item[] = [
-		{ type: "link", text: "Applicants", href: "/admin/applicants" },
+		{ type: "link", text: "Participants", href: "/admin/participants" },
 		{ type: "divider" },
 		{ type: "link", text: "Back to main site", href: "/" },
 	];
+
+	if (isApplicationManager(role)) {
+		navigationItems.unshift(
+			{
+				type: "link",
+				text: "Applicants",
+				href: "/admin/applicants",
+			},
+			{ type: "divider" },
+		);
+	}
 
 	return (
 		<SideNavigation
