+++ apps/site/src/app/admin/participants/Participants.tsx
@@ -1,0 +1,5 @@
+function Participants() {
+	return <></>;
+}
+
+export default Participants;
