+++ apps/site/src/app/admin/participants/page.tsx
@@ -1,0 +1,1 @@
+export { default as default } from "./Participants";
