--- apps/site/src/lib/admin/adminRole.ts
@@ -1,5 +1,0 @@
-const ADMIN_ROLES = ["director", "reviewer"];
-
-export function isAdminRole(role: string | null) {
-	return role !== null && ADMIN_ROLES.includes(role);
-}
